+++ src/olmo_core/train/checkpoint.py
@@ -1,4 +1,5 @@
 import json
+import logging
 import os
 import re
 import tempfile
@@ -34,6 +35,26 @@ from ..utils import wait_for
 from ..version import VERSION
 from .train_module import TrainModule
 
+log = logging.getLogger(__name__)
+
+
+@dataclass
+class CheckpointerConfig(Config):
+    """
+    A configuration class for building :class:`Checkpointer` instances.
+    """
+
+    work_dir: Optional[str] = None
+    save_overwrite: Optional[bool] = None
+    pre_download: bool = False
+
+    def build(self, process_group: Optional[dist.ProcessGroup] = None, **kwargs) -> "Checkpointer":
+        kwargs = {**self.as_dict(exclude_none=True, recurse=False), **kwargs}
+        work_dir = kwargs.pop("work_dir", None)
+        if work_dir is None:
+            raise OLMoConfigurationError("'work_dir' must be provided to build a Checkpointer")
+        return Checkpointer(work_dir=Path(work_dir), process_group=process_group, **kwargs)
+
 
 @dataclass
 class CheckpointMetadata(Config):
@@ -49,8 +70,15 @@ class Checkpointer:
     METADATA_FNAME: ClassVar[str] = ".metadata.json"
     CHECKPOINT_DIR: ClassVar[str] = "step{step}"
 
+    work_dir: Path
     save_overwrite: bool = False
+    pre_download: bool = False
     process_group: Optional[dist.ProcessGroup] = None
+
+    def __post_init__(self):
+        self.work_dir = Path(self.work_dir)
+        if get_fs_local_rank() == 0:
+            self.work_dir.mkdir(exist_ok=True, parents=True)
 
     def save(self, dir: PathOrStr, train_module: TrainModule, train_state: Dict[str, Any]):
         """
@@ -141,6 +169,8 @@ class Checkpointer:
             train_module_dir,
             state_dict,
             process_group=self.process_group,
+            pre_download=is_url(dir) and self.pre_download,
+            work_dir=self.work_dir,
         )
         train_module.load_state_dict(state_dict)
 
@@ -294,7 +324,7 @@ class Checkpointer:
         # Prepare temporary directory.
         tmp_dir: Path
         if is_url(dir):
-            tmp_dir = Path(tempfile.mkdtemp())
+            tmp_dir = Path(tempfile.mkdtemp(dir=str(self.work_dir)))
         else:
             tmp_dir = Path(dir).with_name(Path(dir).name + "-tmp")
             if get_fs_local_rank() == 0: