+++ CHANGELOG.md
@@ -16,6 +16,11 @@ This major release introduces a few breaking changes. As such, we've provided an
 - Added `namespace` argument to `Trainer.record_metric()`.
 - Added `io.copy_dir()` function.
 - Added new LR schedulers: `LinearWithWarmup`, `InvSqrtWithWarmup`, `ConstantWithWarmup`, `SequentialScheduler`.
+- Added option to pre-download checkpoint files from remote storage before trying to load a checkpoint.
+
+### Changed
+
+- Changed storage of shared shard state in sharded checkpoints from smallest shard to lowest rank (normally 0).
 
 ### Changed